requirements.txt
Flask==0.12.2
Pillow==4.2.0
zappa==0.42.2
